Move all types to the types folder
