+++ client/src/api/auth.api.ts
@@ -5,6 +5,9 @@ const apiRefresh = axios.create({
   baseURL: import.meta.env.VITE_API_URL as string,
 });
 
+// Authenticate a registered user into the system
+// @user: username: string, password: string
+// @response: token and refresh token
 export const loginUser = async (user: UserBody) => {
   try {
     const response = await apiRefresh.post("api/v1/token/", user);
@@ -14,6 +17,9 @@ export const loginUser = async (user: UserBody) => {
   }
 };
 
+// Refresh expired token with the provided refresh token
+// @refresh: JWT to refresh token
+// @response: JWT to authenticate user
 export const refreshToken = async (refresh: string) => {
   try {
     const response = await apiRefresh.post("api/v1/token/refresh/", {
@@ -25,6 +31,9 @@ export const refreshToken = async (refresh: string) => {
   }
 };
 
+// Register a new user with a username and a password. The username (email) must be whitelisted before registering
+// @user: username: string, password: string
+// @response: status 201 if created and an email is sent to the email
 export const registerUser = async (user: UserBody) => {
   try {
     const response = await apiRefresh.post("api/v1/user/register/", user);
@@ -34,6 +43,8 @@ export const registerUser = async (user: UserBody) => {
   }
 };
 
+// Activate account by by clicking on a link
+// @activation: uique query parameters to activate account of users 
 export const activateUser = async (activation: ActivationBody) => {
   try {
     const response = await apiRefresh.post("api/v1/user/activate/", activation);
