+++ client/src/components/footer/footer.component.tsx
