--- client/src/components/general/snackbar/types/snackbar.types.ts
@@ -1,6 +1,0 @@
-export type SnackbarParams = {
-  message: string;
-  type: string;
-  open: boolean;
-  handleClose: (event: React.SyntheticEvent | Event, reason?: string) => void;
-};
