+++ client/src/components/sidebar/sidebar.component.tsx
@@ -9,7 +9,7 @@ import {
 } from "@mui/material";
 import WavingHandIcon from "@mui/icons-material/WavingHand";
 import { menu_items } from "./helpers/sidebar.helpers";
-import { MenuItem, SidebarParams } from "./types/sidebar.types";
+import { MenuItem, SidebarParams } from "../../types/types/sidebar.types";
 import ChatBubbleOutlineIcon from "@mui/icons-material/ChatBubbleOutline";
 import {
   Sidebar as SidebarIcon,
