+++ client/src/types/create-dialog/creation-dialog.types.ts
@@ -1,5 +1,5 @@
-import { LabelDTO } from "../../../../types/chatbot/chatbot.types";
-import { TagDTO } from "../../../pages/types/pages.types";
+import { LabelDTO } from "../chatbot/chatbot.types";
+import { TagDTO } from "../page/page.types";
 
 export type CreationDialogParams = {
   open: boolean;
