+++ client/src/types/general/general.types.ts
@@ -2,3 +2,11 @@ export type InputError = {
     error: boolean;
     errorMessage: string;
 }
+
+export type SnackbarParams = {
+    message: string;
+    type: string;
+    open: boolean;
+    handleClose: (event: React.SyntheticEvent | Event, reason?: string) => void;
+  };
+
